make colors a concern
